+++ client/README.md
@@ -3,12 +3,13 @@
 Create a project using this example:
 
 ```bash
-npx thirdweb create --template vite-typescript-starter
+npx thirdweb create --template vite-javascript-starter
 ```
 
-You can start editing the page by modifying `src/main.tsx`. The page auto-updates as you edit the file.
+You can start editing the page by modifying `src/main.jsx`. The page auto-updates as you edit the file.
 
-On `src/index.tsx`, you'll find our `ThirdwebProvider` wrapping your app, this is necessary for our [hooks](https://portal.thirdweb.com/react) and
+On `src/index.jsx`, you'll find our `ThirdwebProvider` wrapping your app,
+this is necessary for our [hooks](https://portal.thirdweb.com/react) and
 [UI Components](https://portal.thirdweb.com/ui-components) to work.
 
 ## Environment Variables
@@ -28,7 +29,7 @@ yarn deploy
 To learn more about thirdweb, Vite and React, take a look at the following resources:
 
 - [thirdweb React Documentation](https://docs.thirdweb.com/react) - learn about our React SDK.
-- [thirdweb TypeScript Documentation](https://docs.thirdweb.com/react) - learn about our JavaScript/TypeScript SDK.
+- [thirdweb JavaScript Documentation](https://docs.thirdweb.com/react) - learn about our JavaScript/TypeScript SDK.
 - [thirdweb Portal](https://docs.thirdweb.com/react) - check our guides and development resources.
 - [Vite Documentation](https://vitejs.dev/guide/) - learn about Vite features.
 - [React documentation](https://reactjs.org/) - learn React.
