+++ client/package.json
@@ -4,23 +4,23 @@
   "version": "0.0.0",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "vite build",
     "preview": "vite preview",
     "deploy": "yarn build && npx thirdweb@latest upload dist"
   },
   "dependencies": {
-    "@thirdweb-dev/react": "^4",
-    "@thirdweb-dev/sdk": "^4",
-    "ethers": "^5",
+    "@thirdweb-dev/react": "^4.4.7",
+    "@thirdweb-dev/sdk": "^4.0.34",
+    "ethers": "^5.7.2",
     "react": "^18.2",
     "react-dom": "^18.2",
     "react-router-dom": "^6.22.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.5",
     "@vitejs/plugin-react": "^2",
-    "typescript": "^4.7.2",
+    "autoprefixer": "^10.4.17",
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.4.1",
     "vite": "^3",
     "vite-plugin-node-polyfills": "^0.7.0"
   }
