+++ client/src/App.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const App = () => {
+  return (
+    <div>App</div>
+  )
+}
+
+export default App
