+++ client/src/assets/index.js
@@ -1,0 +1,31 @@
+import createCampaign from './create-campaign.svg';
+import dashboard from './dashboard.svg';
+import logo from './logo.svg';
+import logout from './logout.svg';
+import payment from './payment.svg';
+import profile from './profile.svg';
+import sun from './sun.svg';
+import withdraw from './withdraw.svg';
+import tagType from './type.svg';
+import search from './search.svg';
+import menu from './menu.svg';
+import money from './money.svg';
+import loader from './loader.svg';
+import thirdweb from './thirdweb.png';
+
+export {
+  tagType,
+  createCampaign,
+  dashboard,
+  logo,
+  logout,
+  payment,
+  profile,
+  sun,
+  withdraw,
+  search,
+  menu,
+  money,
+  loader,
+  thirdweb,
+};
