+++ client/src/main.jsx
@@ -1,0 +1,22 @@
+import { ThirdwebProvider } from "@thirdweb-dev/react";
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { BrowserRouter as Router } from "react-router-dom";
+
+import App from "./App";
+import "./index.css";
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
+  <React.StrictMode>
+    <ThirdwebProvider
+      clientId={import.meta.env.VITE_TEMPLATE_CLIENT_ID}
+      activeChain={Sepolia}
+    >
+      <Router>
+        <App />
+      </Router>
+    </ThirdwebProvider>
+  </React.StrictMode>
+);
