+++ client/tailwind.config.js
@@ -1,0 +1,18 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
+  theme: {
+    extend: {
+      fontFamily: {
+        epilogue: ['Epilogue', 'sans-serif'],
+      },
+      boxShadow: {
+        secondary: '10px 10px 20px rgba(2, 2, 2, 0.25)',
+      },
+    },
+  },
+  plugins: [],
+}
